Use Markdown for Page Section Contents
The styling of each section on the page is preserved, but the body text
of each is stored in a markdown file within `_includes/` and included in
the HTML. In the `_layouts/default.html` file that defines the page, the
markdown source file is specified as an argument for the `include`
statement with key `src`.

This will let the page sections be re-used in other pages that have
different body text.