Create Layout for Second Page
This adds a `title_image` layout for pages whose titles are shown over
an image. We also create a new navbar for non-home pages whose links are
specified in the config YAML. The old navbar is now called `nav_home` as
it is really only useful on the home page.